OAUTH2-210 SF - deduplicate code into Base class
