+++ modules/apps/oauth2-provider/oauth2-provider-rest/src/main/java/com/liferay/oauth2/provider/rest/internal/endpoint/authorize/OAuthAuthorizationDataMessageBodyWriter.java
@@ -14,39 +14,19 @@
 
 package com.liferay.oauth2.provider.rest.internal.endpoint.authorize;
 
-import com.liferay.oauth2.provider.rest.internal.endpoint.authorize.configuration.AuthorizeScreenConfiguration;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.module.configuration.ConfigurationException;
-import com.liferay.portal.kernel.module.configuration.ConfigurationProvider;
-import com.liferay.portal.kernel.settings.CompanyServiceSettingsLocator;
 import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.Http;
-import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.Props;
 import com.liferay.portal.kernel.util.PropsKeys;
 import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.kernel.util.Validator;
-
-import java.io.IOException;
-import java.io.OutputStream;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Type;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
 import javax.ws.rs.Produces;
-import javax.ws.rs.WebApplicationException;
-import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.MultivaluedMap;
-import javax.ws.rs.core.Response;
 import javax.ws.rs.ext.MessageBodyWriter;
 import javax.ws.rs.ext.Provider;
 
-import org.apache.cxf.jaxrs.ext.MessageContext;
 import org.apache.cxf.rs.security.oauth2.common.OAuthAuthorizationData;
 import org.apache.cxf.rs.security.oauth2.utils.OAuthConstants;
 
@@ -62,20 +42,13 @@ import org.osgi.service.component.annotations.Reference;
 		"osgi.jaxrs.application.select=(osgi.jaxrs.name=Liferay.OAuth2.Application)",
 		"osgi.jaxrs.extension=true",
 		"osgi.jaxrs.name=OAuthAuthorizationDataMessageBodyWriter"
-	}
+	},
+	service = MessageBodyWriter.class
 )
 @Produces("text/html")
 @Provider
 public class OAuthAuthorizationDataMessageBodyWriter
-	implements MessageBodyWriter<OAuthAuthorizationData> {
-
-	@Override
-	public long getSize(
-		OAuthAuthorizationData oAuthAuthorizationData, Class<?> clazz,
-		Type genericType, Annotation[] annotations, MediaType mediaType) {
-
-		return -1L;
-	}
+	extends BaseMessageBodyWriter<OAuthAuthorizationData> {
 
 	@Override
 	public boolean isWriteable(
@@ -92,37 +65,17 @@ public class OAuthAuthorizationDataMessageBodyWriter
 		return false;
 	}
 
+	@Activate
+	protected void activate() {
+		_invokerFilterURIMaxLength = GetterUtil.getInteger(
+			_props.get(PropsKeys.INVOKER_FILTER_URI_MAX_LENGTH),
+			_invokerFilterURIMaxLength);
+	}
+
 	@Override
-	public void writeTo(
-			OAuthAuthorizationData oAuthAuthorizationData, Class<?> clazz,
-			Type genericType, Annotation[] annotations, MediaType mediaType,
-			MultivaluedMap<String, Object> httpHeaders,
-			OutputStream entityStream)
-		throws WebApplicationException {
-
-		HttpServletRequest httpServletRequest =
-			_messageContext.getHttpServletRequest();
-
-		String authorizeScreenURL = null;
-
-		try {
-			authorizeScreenURL = getAuthorizeScreenURL(
-				_portal.getCompanyId(httpServletRequest));
-		}
-		catch (ConfigurationException ce) {
-			_log.error("Unable to get authorize screen configuration", ce);
-
-			throw new WebApplicationException(
-				Response.status(
-					Response.Status.INTERNAL_SERVER_ERROR
-				).build());
-		}
-
-		if (!_http.hasDomain(authorizeScreenURL)) {
-			String portalURL = _portal.getPortalURL(httpServletRequest);
-
-			authorizeScreenURL = portalURL + authorizeScreenURL;
-		}
+	protected String writeTo(
+		OAuthAuthorizationData oAuthAuthorizationData,
+		String authorizeScreenURL) {
 
 		authorizeScreenURL = setParameter(
 			authorizeScreenURL, OAuthConstants.AUTHORIZATION_CODE_CHALLENGE,
@@ -160,66 +113,10 @@ public class OAuthAuthorizationDataMessageBodyWriter
 				authorizeScreenURL, OAuthConstants.SCOPE);
 		}
 
-		_messageContext.put("http.request.redirected", Boolean.TRUE);
-
-		HttpServletResponse httpServletResponse =
-			_messageContext.getHttpServletResponse();
-
-		try {
-			httpServletResponse.sendRedirect(authorizeScreenURL);
-		}
-		catch (IOException ioe) {
-			throw new WebApplicationException(ioe);
-		}
+		return authorizeScreenURL;
 	}
-
-	@Activate
-	protected void activate() {
-		_invokerFilterURIMaxLength = GetterUtil.getInteger(
-			_props.get(PropsKeys.INVOKER_FILTER_URI_MAX_LENGTH),
-			_invokerFilterURIMaxLength);
-	}
-
-	protected String getAuthorizeScreenURL(long companyId)
-		throws ConfigurationException {
-
-		AuthorizeScreenConfiguration authorizeScreenConfiguration =
-			_configurationProvider.getConfiguration(
-				AuthorizeScreenConfiguration.class,
-				new CompanyServiceSettingsLocator(
-					companyId, AuthorizeScreenConfiguration.class.getName()));
-
-		return authorizeScreenConfiguration.authorizeScreenURL();
-	}
-
-	protected String removeParameter(String url, String name) {
-		return _http.removeParameter(url, "oauth2_" + name);
-	}
-
-	protected String setParameter(String url, String name, String value) {
-		if (Validator.isBlank(value)) {
-			return url;
-		}
-
-		return _http.addParameter(url, "oauth2_" + name, value);
-	}
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		OAuthAuthorizationDataMessageBodyWriter.class);
-
-	@Reference
-	private ConfigurationProvider _configurationProvider;
-
-	@Reference
-	private Http _http;
 
 	private int _invokerFilterURIMaxLength = 4000;
-
-	@Context
-	private MessageContext _messageContext;
-
-	@Reference
-	private Portal _portal;
 
 	@Reference
 	private Props _props;
